refactor: Genre를 룸 패키지로 이동
