+++ backend/src/main/java/com/digginroom/digginroom/admin/controller/UploadController.java
@@ -2,7 +2,7 @@ package com.digginroom.digginroom.admin.controller;
 
 import com.digginroom.digginroom.admin.controller.dto.UploadRequest;
 import com.digginroom.digginroom.admin.service.UploadService;
-import com.digginroom.digginroom.domain.Genre;
+import com.digginroom.digginroom.domain.room.Genre;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
