+++ backend/src/main/java/com/digginroom/digginroom/domain/recommend/GenreRecommender.java
@@ -1,6 +1,6 @@
 package com.digginroom.digginroom.domain.recommend;
 
-import com.digginroom.digginroom.domain.Genre;
+import com.digginroom.digginroom.domain.room.Genre;
 
 public interface GenreRecommender {
