+++ backend/src/main/java/com/digginroom/digginroom/service/dto/FavoriteGenresRequest.java
@@ -1,6 +1,6 @@
 package com.digginroom.digginroom.service.dto;
 
-import com.digginroom.digginroom.domain.Genre;
+import com.digginroom.digginroom.domain.room.Genre;
 import java.util.List;
 
 public record FavoriteGenresRequest(List<Genre> favoriteGenres) {
